symbolic driver: less than n-1 Hamiltonian terms supported
The option -sx0,x1,x2,...
will tell the symbolic driver that
(1) the Hamiltonian connections x0,x1,x2,... should be enabled
(2) the Hamiltonian connections not listed should be disabled.
Connection labeling depends on the geometry.
For a binary d=1 MERA, x0=0 is the connection between site 0 and 1,
and the i-th connection is between site i and i+1.
x_i should be less than the maximum number of connections or
the code will throw.
If -s is not given it assumes that all Hamiltonian connections
to nearest neighbors are enabled.